% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{end.of.last.month.date}
\alias{end.of.last.month.date}
\title{end.of.last.month.date()}
\usage{
\method{end}{of.last.month.date}()
}
\description{
This function generates the last day of the prior month as a Date object
}
\examples{
end.of.last.month.date()
}

