% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{end.of.last.month.str}
\alias{end.of.last.month.str}
\title{end.of.last.month.str()}
\usage{
\method{end}{of.last.month.str}()
}
\description{
This function generates the last day of the prior month as a character string
}
\examples{
end.of.last.month.str()
}

