% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{end.of.last.week.date}
\alias{end.of.last.week.date}
\title{end.of.last.week.date()}
\usage{
\method{end}{of.last.week.date}()
}
\description{
This function generates the date of the Sunday of last week as a date object
}
\examples{
end.of.last.week.date()
}

