% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{end.of.last.week.str}
\alias{end.of.last.week.str}
\title{end.of.last.week.str()}
\usage{
\method{end}{of.last.week.str}()
}
\description{
This function generates the date of the Sunday of last week as a character string
}
\examples{
end.of.last.week.str()
}

