% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{end.of.last.year.date}
\alias{end.of.last.year.date}
\title{end.of.last.year.date()}
\usage{
\method{end}{of.last.year.date}()
}
\description{
This function generates Dec. 31 of prior year as a date object.
}
\examples{
end.of.last.year.date()
}

