% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{end.of.last.year.str}
\alias{end.of.last.year.str}
\title{end.of.last.year.str()}
\usage{
\method{end}{of.last.year.str}()
}
\description{
This function generates Dec. 31 of prior year as a character string.
}
\examples{
end.of.last.year.str()
}

