% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{end.of.month.from.date}
\alias{end.of.month.from.date}
\title{end.of.month.from.date(mydate)}
\usage{
\method{end}{of.month.from.date}(mydate)
}
\arguments{
\item{mydate}{A date as character string in yyyy-mm-dd format}
}
\description{
This function generates the last day of a month from a character string and returns a character string
}
\examples{
end.of.month.from.date("2015-05-01")
}

