% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{end.of.month.from.datestring}
\alias{end.of.month.from.datestring}
\title{end.of.month.from.datestring(mydate)}
\usage{
\method{end}{of.month.from.datestring}(mydate)
}
\arguments{
\item{mydate}{A date as character string in yyyy-mm-dd format}
}
\description{
This function generates the last day of a month from a character string in format yyyy-mm-dd and returns a character string with the end of that month
}
\examples{
end.of.month.from.date("2015-05-01")
}

