% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{first.of.last.year.date}
\alias{first.of.last.year.date}
\title{first.of.last.year.date()}
\usage{
first.of.last.year.date()
}
\description{
This function generates Jan. 1 of last year as a date object.
}
\examples{
first.of.last.year.date()
}

