% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{first.of.last.year.str}
\alias{first.of.last.year.str}
\title{first.of.last.year.str()}
\usage{
first.of.last.year.str()
}
\description{
This function generates Jan. 1 of last year as a character string.
}
\examples{
first.of.last.year.str()
}

