% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{first.of.month.date}
\alias{first.of.month.date}
\title{first.of.month.date()}
\usage{
first.of.month.date()
}
\description{
This function generates the 1st of the current month as a Date object
}
\examples{
first.of.month.date()
}

