% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{first.of.month.str}
\alias{first.of.month.str}
\title{first.of.month.str()}
\usage{
first.of.month.str()
}
\description{
This function generates the 1st of the current month as a character string
}
\examples{
first.of.month.str()
}

