% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{start.of.last.week.date}
\alias{start.of.last.week.date}
\title{start.of.last.week.date()}
\usage{
\method{start}{of.last.week.date}()
}
\description{
This function generates the date of the Monday of last week as a date object
}
\examples{
start.of.last.week.date()
}

