% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{start.of.last.week.str}
\alias{start.of.last.week.str}
\title{start.of.last.week.str()}
\usage{
\method{start}{of.last.week.str}()
}
\description{
This function generates the date of the Monday of last week as a character string
}
\examples{
start.of.last.week.str()
}

