% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{start.of.this.week.date}
\alias{start.of.this.week.date}
\title{start.of.this.week.date()}
\usage{
\method{start}{of.this.week.date}()
}
\description{
This function generates the date of Monday of this week as a date object
}
\examples{
start.of.this.week.date()
}

