% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{start.of.this.week.str}
\alias{start.of.this.week.str}
\title{start.of.this.week.str()}
\usage{
\method{start}{of.this.week.str}()
}
\description{
This function generates the date of Monday of this week as a character string
}
\examples{
start.of.this.week.str()
}

