% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{today.date}
\alias{today.date}
\title{today.date()}
\usage{
today.date()
}
\description{
This function generates a Date object with the current date. It is a wrapper for Sys.Date() to keep rdateutils function syntax.
}
\examples{
today.date()
}

