% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{yesterday.date}
\alias{yesterday.date}
\title{yesterday.date()}
\usage{
yesterday.date()
}
\description{
This function generates a Date object with yesterday's date.
}
\examples{
yesterday.date()
}

