% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdateutils.R
\name{yesterday.str}
\alias{yesterday.str}
\title{yesterday.str()}
\usage{
yesterday.str()
}
\description{
This function generates a character string with yesterday's date.
}
\examples{
yesterday.str()
}

